+++ packages/commons/src/events/index.ts
@@ -36,7 +36,6 @@ export * from './utils'
 export * from './defineConfig'
 export * from './DeduplicationConfig'
 export * from './transactions'
-export * from './User'
 export * from './FieldType'
 export * from './ActionType'
 export * from './FieldTypeMapping'
