+++ packages/commons/src/events/serializers/user/serializer.ts
@@ -9,7 +9,7 @@
  * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
  */
 
-import { User } from '../../User'
+import { User } from 'src/users/User'
 import { z } from 'zod'
 
 export const SerializedUserField = z.object({
