+++ packages/commons/src/index.ts
@@ -17,8 +17,7 @@ export * from './logger'
 export * from './roles'
 export * from './search'
 export * from './events'
-export * from './users/service'
-export * from './users/systemRole'
+export * from './users/User'
 export * from './authentication'
 export * from './utils'
 export * from './countryconfig'
